Calculo de rutas, pendiente triagulación con driver

+++ src/main/java/FormatoBase/proyectoJWT/model/entity/Driver.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -37,7 +38,7 @@ public class Driver extends BaseEntity {
     private Double longitud;
 
     @Column(name = "costo_activacion")
-    private float costoActivacion;
+    private BigDecimal costoActivacion;
 
     @JsonIgnore
     @ManyToOne(cascade = CascadeType.ALL, fetch = FetchType.LAZY)
